requirements.txt
# Core dependencies
pathway==0.13.0
pandas==2.1.4
numpy==1.26.2

# LLM and embeddings
anthropic==0.34.0
openai==1.35.0
sentence-transformers==2.7.0

# Vector store
chromadb==0.4.22
faiss-cpu==1.7.4

# NLP utilities
nltk==3.8.1
spacy==3.7.2
transformers==4.36.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0